Move the context to the folder contexts
