+++ src/pages/report-payouts/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { ConditionProvider } from './Condition.context'
+import { ConditionProvider } from '../../contexts/ConditionContext'
 import { ReportPayouts } from './ReportPayouts'
 
 export const ReportPayoutsContainer: React.FC = () => {
